@@ -3,14 +3,14 @@
 #
 
 name: 'p2pquake_api_v2'
-version: '0.1.4'
+version: '1.0.0'
 description: 'OpenAPI API client'
 homepage: 'homepage'
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
-  http: ^0.13.4
-  intl: ^0.17.0
-  meta: '^1.1.8'
+  http: ^0.13.6
+  intl: ^0.18.1
+  meta: ^1.16.0
 dev_dependencies:
-  test: '^1.3.0'
+  test: ^1.24.3
