@@ -1,7 +1,6 @@
 //
 // AUTO-GENERATED FILE, DO NOT MODIFY!
 //
-// @dart=2.0
 
 // ignore_for_file: unused_element, unused_import
 // ignore_for_file: always_put_required_named_parameters_first
@@ -10,7 +9,7 @@
 part of p2pquake_v2_api.api;
 
 class JsonApi {
-  JsonApi([ApiClient apiClient]) : apiClient = apiClient ?? defaultApiClient;
+  JsonApi([ApiClient? apiClient]) : apiClient = apiClient ?? defaultApiClient;
 
   final ApiClient apiClient;
 
@@ -55,85 +54,65 @@ class JsonApi {
   /// * [List<Object>] prefectures:
   ///   各都道府県の最低震度。 \"兵庫県,10\" のように指定します。
   Future<Response> jmaQuakeGetWithHttpInfo(
-      {int limit,
-      int offset,
-      int order,
-      String sinceDate,
-      String untilDate,
-      String quakeType,
-      num minMagnitude,
-      num maxMagnitude,
-      int minScale,
-      int maxScale,
-      List<Object> prefectures}) async {
+      {int? limit,
+      int? offset,
+      int? order,
+      String? sinceDate,
+      String? untilDate,
+      String? quakeType,
+      num? minMagnitude,
+      num? maxMagnitude,
+      int? minScale,
+      int? maxScale,
+      List<Object>? prefectures}) async {
     // Verify required params are set.
 
     final path = '/jma/quake'.replaceAll('{format}', 'json');
 
-    Object postBody;
+    Object? postBody;
 
     final queryParams = <QueryParam>[];
     final headerParams = <String, String>{};
     final formParams = <String, String>{};
 
     if (limit != null) {
-      queryParams
-          .addAll(_convertParametersForCollectionFormat('', 'limit', limit));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'limit', limit));
     }
     if (offset != null) {
-      queryParams
-          .addAll(_convertParametersForCollectionFormat('', 'offset', offset));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'offset', offset));
     }
     if (order != null) {
-      queryParams
-          .addAll(_convertParametersForCollectionFormat('', 'order', order));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'order', order));
     }
     if (sinceDate != null) {
-      queryParams.addAll(
-          _convertParametersForCollectionFormat('', 'since_date', sinceDate));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'since_date', sinceDate));
     }
     if (untilDate != null) {
-      queryParams.addAll(
-          _convertParametersForCollectionFormat('', 'until_date', untilDate));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'until_date', untilDate));
     }
     if (quakeType != null) {
-      queryParams.addAll(
-          _convertParametersForCollectionFormat('', 'quake_type', quakeType));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'quake_type', quakeType));
     }
     if (minMagnitude != null) {
-      queryParams.addAll(_convertParametersForCollectionFormat(
-          '', 'min_magnitude', minMagnitude));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'min_magnitude', minMagnitude));
     }
     if (maxMagnitude != null) {
-      queryParams.addAll(_convertParametersForCollectionFormat(
-          '', 'max_magnitude', maxMagnitude));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'max_magnitude', maxMagnitude));
     }
     if (minScale != null) {
-      queryParams.addAll(
-          _convertParametersForCollectionFormat('', 'min_scale', minScale));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'min_scale', minScale));
     }
     if (maxScale != null) {
-      queryParams.addAll(
-          _convertParametersForCollectionFormat('', 'max_scale', maxScale));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'max_scale', maxScale));
     }
     if (prefectures != null) {
-      queryParams.addAll(_convertParametersForCollectionFormat(
-          'multi', 'prefectures[]', prefectures));
+      queryParams
+          .addAll(_convertParametersForCollectionFormat('multi', 'prefectures[]', prefectures));
     }
 
     final contentTypes = <String>[];
-    final nullableContentType =
-        contentTypes.isNotEmpty ? contentTypes[0] : null;
+    final nullableContentType = contentTypes.isNotEmpty ? contentTypes[0] : null;
     final authNames = <String>[];
-
-    if (nullableContentType != null &&
-        nullableContentType.toLowerCase().startsWith('multipart/form-data')) {
-      bool hasFields = false;
-      final mp = MultipartRequest(null, null);
-      if (hasFields) {
-        postBody = mp;
-      }
-    } else {}
 
     return await apiClient.invokeAPI(
       path,
@@ -185,18 +164,18 @@ class JsonApi {
   ///
   /// * [List<Object>] prefectures[]:
   ///   各都道府県の最低震度。 \"兵庫県,10\" のように指定します。
-  Future<List<JMAQuake>> jmaQuakeGet(
-      {int limit,
-      int offset,
-      int order,
-      String sinceDate,
-      String untilDate,
-      String quakeType,
-      num minMagnitude,
-      num maxMagnitude,
-      int minScale,
-      int maxScale,
-      List<Object> prefectures}) async {
+  Future<List<JMAQuake>?> jmaQuakeGet(
+      {int? limit,
+      int? offset,
+      int? order,
+      String? sinceDate,
+      String? untilDate,
+      String? quakeType,
+      num? minMagnitude,
+      num? maxMagnitude,
+      int? minScale,
+      int? maxScale,
+      List<Object>? prefectures}) async {
     final response = await jmaQuakeGetWithHttpInfo(
         limit: limit,
         offset: offset,
@@ -210,14 +189,13 @@ class JsonApi {
         maxScale: maxScale,
         prefectures: prefectures);
     if (response.statusCode >= HttpStatus.badRequest) {
-      throw ApiException(response.statusCode, _decodeBodyBytes(response));
+      throw ApiException(response.statusCode, _decodeBodyBytes(response) ?? "");
     }
     // When a remote server returns no body with a status of 204, we shall not decode it.
     // At the time of writing this, `dart:convert` will throw an "Unexpected end of input"
     // FormatException when trying to decode an empty string.
-    if (response.body != null && response.statusCode != HttpStatus.noContent) {
-      return (apiClient.deserialize(
-              _decodeBodyBytes(response), 'List<JMAQuake>') as List)
+    if (response.body != "" && response.statusCode != HttpStatus.noContent) {
+      return (apiClient.deserialize(_decodeBodyBytes(response)!, 'List<JMAQuake>') as List)
           .map((item) => item as JMAQuake)
           .toList(growable: false);
     }
@@ -242,25 +220,15 @@ class JsonApi {
         .replaceAll('{format}', 'json')
         .replaceAll('{' + 'id' + '}', id.toString());
 
-    Object postBody;
+    Object? postBody;
 
     final queryParams = <QueryParam>[];
     final headerParams = <String, String>{};
     final formParams = <String, String>{};
 
     final contentTypes = <String>[];
-    final nullableContentType =
-        contentTypes.isNotEmpty ? contentTypes[0] : null;
+    final nullableContentType = contentTypes.isNotEmpty ? contentTypes[0] : null;
     final authNames = <String>[];
-
-    if (nullableContentType != null &&
-        nullableContentType.toLowerCase().startsWith('multipart/form-data')) {
-      bool hasFields = false;
-      final mp = MultipartRequest(null, null);
-      if (hasFields) {
-        postBody = mp;
-      }
-    } else {}
 
     return await apiClient.invokeAPI(
       path,
@@ -280,17 +248,16 @@ class JsonApi {
   ///
   /// * [String] id (required):
   ///   ID
-  Future<JMAQuake> jmaQuakeIdGet(String id) async {
+  Future<JMAQuake?> jmaQuakeIdGet(String id) async {
     final response = await jmaQuakeIdGetWithHttpInfo(id);
     if (response.statusCode >= HttpStatus.badRequest) {
-      throw ApiException(response.statusCode, _decodeBodyBytes(response));
+      throw ApiException(response.statusCode, _decodeBodyBytes(response) ?? "");
     }
     // When a remote server returns no body with a status of 204, we shall not decode it.
     // At the time of writing this, `dart:convert` will throw an "Unexpected end of input"
     // FormatException when trying to decode an empty string.
-    if (response.body != null && response.statusCode != HttpStatus.noContent) {
-      return apiClient.deserialize(_decodeBodyBytes(response), 'JMAQuake')
-          as JMAQuake;
+    if (response.body != "" && response.statusCode != HttpStatus.noContent) {
+      return apiClient.deserialize(_decodeBodyBytes(response)!, 'JMAQuake') as JMAQuake?;
     }
     return null;
   }
@@ -318,55 +285,36 @@ class JsonApi {
   /// * [String] untilDate:
   ///   指定日かそれ以前 (yyyyMMdd形式)
   Future<Response> jmaTsunamiGetWithHttpInfo(
-      {int limit,
-      int offset,
-      int order,
-      String sinceDate,
-      String untilDate}) async {
+      {int? limit, int? offset, int? order, String? sinceDate, String? untilDate}) async {
     // Verify required params are set.
 
     final path = '/jma/tsunami'.replaceAll('{format}', 'json');
 
-    Object postBody;
+    Object? postBody;
 
     final queryParams = <QueryParam>[];
     final headerParams = <String, String>{};
     final formParams = <String, String>{};
 
     if (limit != null) {
-      queryParams
-          .addAll(_convertParametersForCollectionFormat('', 'limit', limit));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'limit', limit));
     }
     if (offset != null) {
-      queryParams
-          .addAll(_convertParametersForCollectionFormat('', 'offset', offset));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'offset', offset));
     }
     if (order != null) {
-      queryParams
-          .addAll(_convertParametersForCollectionFormat('', 'order', order));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'order', order));
     }
     if (sinceDate != null) {
-      queryParams.addAll(
-          _convertParametersForCollectionFormat('', 'since_date', sinceDate));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'since_date', sinceDate));
     }
     if (untilDate != null) {
-      queryParams.addAll(
-          _convertParametersForCollectionFormat('', 'until_date', untilDate));
+      queryParams.addAll(_convertParametersForCollectionFormat('', 'until_date', untilDate));
     }
 
     final contentTypes = <String>[];
-    final nullableContentType =
-        contentTypes.isNotEmpty ? contentTypes[0] : null;
+    final nullableContentType = contentTypes.isNotEmpty ? contentTypes[0] : null;
     final authNames = <String>[];
-
-    if (nullableContentType != null &&
-        nullableContentType.toLowerCase().startsWith('multipart/form-data')) {
-      bool hasFields = false;
-      final mp = MultipartRequest(null, null);
-      if (hasFields) {
-        postBody = mp;
-      }
-    } else {}
 
     return await apiClient.invokeAPI(
       path,
@@ -400,27 +348,18 @@ class JsonApi {
   ///
   /// * [String] untilDate:
   ///   指定日かそれ以前 (yyyyMMdd形式)
-  Future<List<JMATsunami>> jmaTsunamiGet(
-      {int limit,
-      int offset,
-      int order,
-      String sinceDate,
-      String untilDate}) async {
+  Future<List<JMATsunami>?> jmaTsunamiGet(
+      {int? limit, int? offset, int? order, String? sinceDate, String? untilDate}) async {
     final response = await jmaTsunamiGetWithHttpInfo(
-        limit: limit,
-        offset: offset,
-        order: order,
-        sinceDate: sinceDate,
-        untilDate: untilDate);
+        limit: limit, offset: offset, order: order, sinceDate: sinceDate, untilDate: untilDate);
     if (response.statusCode >= HttpStatus.badRequest) {
-      throw ApiException(response.statusCode, _decodeBodyBytes(response));
+      throw ApiException(response.statusCode, _decodeBodyBytes(response) ?? "");
     }
     // When a remote server returns no body with a status of 204, we shall not decode it.
     // At the time of writing this, `dart:convert` will throw an "Unexpected end of input"
     // FormatException when trying to decode an empty string.
-    if (response.body != null && response.statusCode != HttpStatus.noContent) {
-      return (apiClient.deserialize(
-              _decodeBodyBytes(response), 'List<JMATsunami>') as List)
+    if (response.body != "" && response.statusCode != HttpStatus.noContent) {
+      return (apiClient.deserialize(_decodeBodyBytes(response)!, 'List<JMATsunami>') as List)
           .map((item) => item as JMATsunami)
           .toList(growable: false);
     }
@@ -445,25 +384,15 @@ class JsonApi {
         .replaceAll('{format}', 'json')
         .replaceAll('{' + 'id' + '}', id.toString());
 
-    Object postBody;
+    Object? postBody;
 
     final queryParams = <QueryParam>[];
     final headerParams = <String, String>{};
     final formParams = <String, String>{};
 
     final contentTypes = <String>[];
-    final nullableContentType =
-        contentTypes.isNotEmpty ? contentTypes[0] : null;
+    final nullableContentType = contentTypes.isNotEmpty ? contentTypes[0] : null;
     final authNames = <String>[];
-
-    if (nullableContentType != null &&
-        nullableContentType.toLowerCase().startsWith('multipart/form-data')) {
-      bool hasFields = false;
-      final mp = MultipartRequest(null, null);
-      if (hasFields) {
-        postBody = mp;
-      }
-    } else {}
 
     return await apiClient.invokeAPI(
       path,
@@ -483,17 +412,16 @@ class JsonApi {
   ///
   /// * [String] id (required):
   ///   ID
-  Future<JMATsunami> jmaTsunamiIdGet(String id) async {
+  Future<JMATsunami?> jmaTsunamiIdGet(String id) async {
     final response = await jmaTsunamiIdGetWithHttpInfo(id);
     if (response.statusCode >= HttpStatus.badRequest) {
-      throw ApiException(response.statusCode, _decodeBodyBytes(response));
+      throw ApiException(response.statusCode, _decodeBodyBytes(response) ?? "");
     }
     // When a remote server returns no body with a status of 204, we shall not decode it.
     // At the time of writing this, `dart:convert` will throw an "Unexpected end of input"
     // FormatException when trying to decode an empty string.
-    if (response.body != null && response.statusCode != HttpStatus.noContent) {
-      return apiClient.deserialize(_decodeBodyBytes(response), 'JMATsunami')
-          as JMATsunami;
+    if (response.body != "" && response.statusCode != HttpStatus.noContent) {
+      return apiClient.deserialize(_decodeBodyBytes(response)!, 'JMATsunami') as JMATsunami?;
     }
     return null;
   }
