@@ -1,7 +1,7 @@
 //
 // AUTO-GENERATED FILE, DO NOT MODIFY!
 //
-// @dart=2.0
+
 
 // ignore_for_file: unused_element, unused_import
 // ignore_for_file: always_put_required_named_parameters_first
@@ -17,8 +17,8 @@ class ApiException implements Exception {
 
   int code = 0;
   String message;
-  Exception innerException;
-  StackTrace stackTrace;
+  Exception? innerException;
+  StackTrace? stackTrace;
 
   String toString() {
     if (message == null) {
