@@ -1,7 +1,7 @@
 //
 // AUTO-GENERATED FILE, DO NOT MODIFY!
 //
-// @dart=2.0
+
 
 // ignore_for_file: unused_element, unused_import
 // ignore_for_file: always_put_required_named_parameters_first
@@ -15,8 +15,8 @@ class ApiKeyAuth implements Authentication {
   final String location;
   final String paramName;
 
-  String apiKeyPrefix;
-  String apiKey;
+  String? apiKeyPrefix;
+  String? apiKey;
 
   @override
   void applyToParams(
