@@ -1,7 +1,7 @@
 //
 // AUTO-GENERATED FILE, DO NOT MODIFY!
 //
-// @dart=2.0
+
 
 // ignore_for_file: unused_element, unused_import
 // ignore_for_file: always_put_required_named_parameters_first
@@ -10,8 +10,8 @@
 part of p2pquake_v2_api.api;
 
 class HttpBasicAuth implements Authentication {
-  String username;
-  String password;
+  String? username;
+  String? password;
 
   @override
   void applyToParams(
