@@ -1,7 +1,7 @@
 //
 // AUTO-GENERATED FILE, DO NOT MODIFY!
 //
-// @dart=2.0
+
 
 // ignore_for_file: unused_element, unused_import
 // ignore_for_file: always_put_required_named_parameters_first
@@ -12,7 +12,7 @@ part of p2pquake_v2_api.api;
 class OAuth implements Authentication {
   OAuth({this.accessToken});
 
-  String accessToken;
+  String? accessToken;
 
   @override
   void applyToParams(
