@@ -1,7 +1,6 @@
 //
 // AUTO-GENERATED FILE, DO NOT MODIFY!
 //
-// @dart=2.0
 
 // ignore_for_file: unused_element, unused_import
 // ignore_for_file: always_put_required_named_parameters_first
@@ -12,15 +11,15 @@ part of p2pquake_v2_api.api;
 class Areapeers extends BasicData {
   /// Returns a new [Areapeers] instance.
   Areapeers({
-    @required this.id,
-    @required this.code,
-    @required this.time,
+    required this.id,
+    required this.code,
+    required this.time,
     this.areas = const [],
-  });
+  }) : super(id: id, code: code, time: time);
 
   /// Returns a new [Areapeers] instance and optionally import its values from
   /// [json] if it's non-null.
-  Areapeers.fromJson(Map<String, dynamic> json) {
+  Areapeers.fromJson(Map<String, dynamic>? json) : super.fromJson(json) {
     if (json != null) {
       id = json['id'];
       code = json['code'];
@@ -30,16 +29,16 @@ class Areapeers extends BasicData {
   }
 
   /// 情報を一意に識別するID
-  String id;
+  String? id;
 
   /// 情報コード。常に555です。
-  int code;
+  int? code;
 
   /// 受信日時。形式は `2006/01/02 15:04:05.999` です。
-  String time;
+  String? time;
 
   /// ピアの地域分布
-  List<AreapeersAllOfAreas> areas;
+  List<AreapeersAllOfAreas>? areas;
 
   @override
   bool operator ==(Object other) =>
@@ -51,12 +50,10 @@ class Areapeers extends BasicData {
           other.areas == areas;
 
   @override
-  int get hashCode =>
-      id.hashCode + (code?.hashCode ?? 0) + time.hashCode + areas.hashCode;
+  int get hashCode => id.hashCode + (code?.hashCode ?? 0) + time.hashCode + areas.hashCode;
 
   @override
-  String toString() =>
-      'Areapeers[id=$id, code=$code, time=$time, areas=$areas]';
+  String toString() => 'Areapeers[id=$id, code=$code, time=$time, areas=$areas]';
 
   Map<String, dynamic> toJson() {
     final json = <String, dynamic>{};
@@ -75,18 +72,16 @@ class Areapeers extends BasicData {
     return json;
   }
 
-  static List<Areapeers> listFromJson(
+  static List<Areapeers>? listFromJson(
     List<dynamic> json, {
-    bool emptyIsNull,
-    bool growable,
+    bool? emptyIsNull,
+    bool? growable,
   }) =>
       json == null || json.isEmpty
           ? true == emptyIsNull
               ? null
               : <Areapeers>[]
-          : json
-              .map((v) => Areapeers.fromJson(v))
-              .toList(growable: true == growable);
+          : json.map((v) => Areapeers.fromJson(v)).toList(growable: true == growable);
 
   static Map<String, Areapeers> mapFromJson(Map<String, dynamic> json) {
     final map = <String, Areapeers>{};
@@ -97,16 +92,15 @@ class Areapeers extends BasicData {
   }
 
   // maps a json object with a list of Areapeers-objects as value to a dart map
-  static Map<String, List<Areapeers>> mapListFromJson(
+  static Map<String, List<Areapeers>?> mapListFromJson(
     Map<String, dynamic> json, {
-    bool emptyIsNull,
-    bool growable,
+    bool? emptyIsNull,
+    bool? growable,
   }) {
-    final map = <String, List<Areapeers>>{};
+    final map = <String, List<Areapeers>?>{};
     if (json != null && json.isNotEmpty) {
       json.forEach((String key, dynamic v) {
-        map[key] = Areapeers.listFromJson(v,
-            emptyIsNull: emptyIsNull, growable: growable);
+        map[key] = Areapeers.listFromJson(v, emptyIsNull: emptyIsNull, growable: growable);
       });
     }
     return map;
