@@ -1,7 +1,6 @@
 //
 // AUTO-GENERATED FILE, DO NOT MODIFY!
 //
-// @dart=2.0
 
 // ignore_for_file: unused_element, unused_import
 // ignore_for_file: always_put_required_named_parameters_first
@@ -12,15 +11,15 @@ part of p2pquake_v2_api.api;
 class Userquake extends BasicData {
   /// Returns a new [Userquake] instance.
   Userquake({
-    @required this.id,
-    @required this.code,
-    @required this.time,
-    @required this.area,
-  });
+    required this.id,
+    required this.code,
+    required this.time,
+    required this.area,
+  }) : super(id: id, code: code, time: time);
 
   /// Returns a new [Userquake] instance and optionally import its values from
   /// [json] if it's non-null.
-  Userquake.fromJson(Map<String, dynamic> json) {
+  Userquake.fromJson(Map<String, dynamic>? json) : super.fromJson(json) {
     if (json != null) {
       id = json['id'];
       code = json['code'];
@@ -30,16 +29,16 @@ class Userquake extends BasicData {
   }
 
   /// 情報を一意に識別するID
-  String id;
+  String? id;
 
   /// 情報コード。常に561です。
-  int code;
+  int? code;
 
   /// 受信日時。形式は `2006/01/02 15:04:05.999` です。
-  String time;
+  String? time;
 
   /// 地域コード
-  int area;
+  int? area;
 
   @override
   bool operator ==(Object other) =>
@@ -51,8 +50,7 @@ class Userquake extends BasicData {
           other.area == area;
 
   @override
-  int get hashCode =>
-      id.hashCode + (code?.hashCode ?? 0) + time.hashCode + area.hashCode;
+  int get hashCode => id.hashCode + (code?.hashCode ?? 0) + time.hashCode + area.hashCode;
 
   @override
   String toString() => 'Userquake[id=$id, code=$code, time=$time, area=$area]';
@@ -74,18 +72,16 @@ class Userquake extends BasicData {
     return json;
   }
 
-  static List<Userquake> listFromJson(
+  static List<Userquake>? listFromJson(
     List<dynamic> json, {
-    bool emptyIsNull,
-    bool growable,
+    bool? emptyIsNull,
+    bool? growable,
   }) =>
       json == null || json.isEmpty
           ? true == emptyIsNull
               ? null
               : <Userquake>[]
-          : json
-              .map((v) => Userquake.fromJson(v))
-              .toList(growable: true == growable);
+          : json.map((v) => Userquake.fromJson(v)).toList(growable: true == growable);
 
   static Map<String, Userquake> mapFromJson(Map<String, dynamic> json) {
     final map = <String, Userquake>{};
@@ -96,16 +92,15 @@ class Userquake extends BasicData {
   }
 
   // maps a json object with a list of Userquake-objects as value to a dart map
-  static Map<String, List<Userquake>> mapListFromJson(
+  static Map<String, List<Userquake>?> mapListFromJson(
     Map<String, dynamic> json, {
-    bool emptyIsNull,
-    bool growable,
+    bool? emptyIsNull,
+    bool? growable,
   }) {
-    final map = <String, List<Userquake>>{};
+    final map = <String, List<Userquake>?>{};
     if (json != null && json.isNotEmpty) {
       json.forEach((String key, dynamic v) {
-        map[key] = Userquake.listFromJson(v,
-            emptyIsNull: emptyIsNull, growable: growable);
+        map[key] = Userquake.listFromJson(v, emptyIsNull: emptyIsNull, growable: growable);
       });
     }
     return map;
