@@ -1,7 +1,7 @@
 //
 // AUTO-GENERATED FILE, DO NOT MODIFY!
 //
-// @dart=2.0
+
 
 // ignore_for_file: unused_element, unused_import
 // ignore_for_file: always_put_required_named_parameters_first
@@ -19,7 +19,7 @@ class UserquakeEvaluationAllOfAreaConfidences {
 
   /// Returns a new [UserquakeEvaluationAllOfAreaConfidences] instance and optionally import its values from
   /// [json] if it's non-null.
-  UserquakeEvaluationAllOfAreaConfidences.fromJson(Map<String, dynamic> json) {
+  UserquakeEvaluationAllOfAreaConfidences.fromJson(Map<String, dynamic>? json) {
     if (json != null) {
       confidence =
           json['confidence'] == null ? null : json['confidence'].toDouble();
@@ -29,13 +29,13 @@ class UserquakeEvaluationAllOfAreaConfidences {
   }
 
   /// 信頼度（0～1） P2P地震情報 Beta3 においては、 0未満: F、0.0以上0.2未満: E、0.2以上0.4未満: D、0.4以上0.6未満: C、0.6以上0.8未満: B、0.8以上: A です。
-  num confidence;
+  num? confidence;
 
   /// 件数
-  int count;
+  int? count;
 
   /// P2P地震情報 Beta3 における信頼度表示
-  String display;
+  String? display;
 
   @override
   bool operator ==(Object other) =>
@@ -66,10 +66,10 @@ class UserquakeEvaluationAllOfAreaConfidences {
     return json;
   }
 
-  static List<UserquakeEvaluationAllOfAreaConfidences> listFromJson(
+  static List<UserquakeEvaluationAllOfAreaConfidences>? listFromJson(
     List<dynamic> json, {
-    bool emptyIsNull,
-    bool growable,
+    bool? emptyIsNull,
+    bool? growable,
   }) =>
       json == null || json.isEmpty
           ? true == emptyIsNull
@@ -80,7 +80,7 @@ class UserquakeEvaluationAllOfAreaConfidences {
               .toList(growable: true == growable);
 
   static Map<String, UserquakeEvaluationAllOfAreaConfidences> mapFromJson(
-      Map<String, dynamic> json) {
+      Map<String, dynamic>? json) {
     final map = <String, UserquakeEvaluationAllOfAreaConfidences>{};
     if (json != null && json.isNotEmpty) {
       json.forEach((String key, dynamic v) =>
@@ -90,13 +90,13 @@ class UserquakeEvaluationAllOfAreaConfidences {
   }
 
   // maps a json object with a list of UserquakeEvaluationAllOfAreaConfidences-objects as value to a dart map
-  static Map<String, List<UserquakeEvaluationAllOfAreaConfidences>>
+  static Map<String, List<UserquakeEvaluationAllOfAreaConfidences>?>
       mapListFromJson(
     Map<String, dynamic> json, {
-    bool emptyIsNull,
-    bool growable,
+    bool? emptyIsNull,
+    bool? growable,
   }) {
-    final map = <String, List<UserquakeEvaluationAllOfAreaConfidences>>{};
+    final map = <String, List<UserquakeEvaluationAllOfAreaConfidences>?>{};
     if (json != null && json.isNotEmpty) {
       json.forEach((String key, dynamic v) {
         map[key] = UserquakeEvaluationAllOfAreaConfidences.listFromJson(v,
